requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.27.0